requirements.txt
category_encoders==2.6.3
fuzzywuzzy==0.18.0
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
Requests==2.32.3
scikit_learn==1.5.0
seaborn==0.13.2
streamlit==1.33.0
